feat: 添加 Blazor 支持及多项功能和样式更新
- 在 `UraDocs.sln` 文件中，添加了 `UraDocs.Frontend` 项目，并更新了相应的配置。
- 在 `FileHelper.cs` 文件中，修改了 `GetMarkDownPath` 方法，去除了路径前缀中的 `.` 和 `/`。
- 在 `Program.cs` 文件中，添加了 Razor Pages 支持，并在应用程序中启用了 Blazor 框架文件和 Razor Pages。
- 在 `launchSettings.json` 文件中，移除了 `launchUrl` 的 `swagger` 配置。
- 在 `DocumentService.cs` 文件中，修改了 `GetFileHashAsync` 和 `GeneratorHtmlAsync` 方法，并添加了 `ReadTextAsync` 方法。
- 在 `FileWaterBackgroundService.cs` 文件中，添加了对临时文件的检查，避免对临时文件进行处理。
- 在 `UraDocs.ApiService.csproj` 文件中，添加了对 `Microsoft.AspNetCore.Components.WebAssembly.Server` 和 `Microsoft.Extensions.Http` 的引用。
- 在 `Program.cs` 文件中，注释掉了 `UraDocs_Web` 项目的引用。
- 在 `UraDocs.AppHost.csproj` 文件中，移除了对 `UraDocs.Web` 项目的引用。
- 在 `App.razor` 文件中，移除了 `InteractiveAuto` 渲染模式。
- 在 `Program.cs` 文件中，注释掉了 `AddServiceDefaults` 方法，并修改了 Razor 组件的配置。
- 在 `UraDocs.Web.csproj` 文件中，移除了对 `UraDocs.ServiceDefaults` 项目的引用。
- 在 `app.css` 文件中，移除了注释。
- 在 `README - 副本.MD` 文件中，添加了 Ura Docs 的描述和功能列表。
- 在多个 HTML 文件中，添加了 Ura Docs 的描述和功能列表。
- 在 `App.razor` 文件中，添加了 `Router` 组件。
- 在 `Program.cs` 文件中，添加了 Blazor WebAssembly 的相关配置。
- 在 `UraDocs.Frontend.csproj` 文件中，添加了 Blazor WebAssembly 项目的配置。
- 在 `_Imports.razor` 文件中，添加了相关的 using 语句。
- 在 `MainLayout.razor` 文件中，添加了布局和主题配置。
- 在 `MainLayout.razor.css` 文件中，添加了布局的样式。
- 在 `NavMenu.razor` 文件中，添加了导航菜单。
- 在 `NavMenu.razor.css` 中添加了多个 CSS 样式类，用于自定义导航菜单的外观和行为。
- 在 `Counter.razor` 中添加了一个计数器页面，包含一个按钮和一个显示当前计数的文本。
- 在 `Document.razor` 中添加了一个文档页面，使用 `MudBlazor` 组件库来显示文档树和文档内容。
- 在 `Home.razor` 中添加了主页内容，包括一个按钮链接到文档页面。
- 添加了 `launchSettings.json` 文件，用于配置项目的启动设置，包括 IIS Express 和 HTTP/HTTPS 配置。
- 添加了 `UraService.cs` 文件，定义了一个服务类，用于从 API 获取菜单数据。
- 添加了 `favicon.png` 和 `icon-192.png` 图标文件。
- 添加了 `index.html` 文件，定义了应用程序的 HTML 结构和引用的样式表、脚本。
- 添加了多个 CSS 样式类，用于自定义页面元素的外观和行为。
- 添加了一个新的 JSON 数组，其中包含五个天气记录。每个记录包括日期、摄氏温度和天气概述。

+++ src/UraDocs/UraDocs.ApiService/Services/FileWaterBackgroundService.cs
@@ -77,8 +77,25 @@ public class FileWaterBackgroundService : BackgroundService
         using var scope = _serviceProvider.CreateScope();
         var documentService = scope.ServiceProvider.GetRequiredService<DocumentService>();
 
-        await documentService.DeleteHtmlAsync(GetRelativePath(e.OldFullPath));
-        await documentService.GeneratorHtmlAsync(GetRelativePath(e.FullPath));
+        if(IsTemporaryFile(e.Name ?? string.Empty))
+        {
+            return;
+        }
+
+        if(IsTemporaryFile(e.OldName ?? string.Empty))
+        {
+            await documentService.UpdateHtmlAsync(GetRelativePath(e.FullPath));
+        }
+        else
+        {
+            await documentService.DeleteHtmlAsync(GetRelativePath(e.OldFullPath));
+            await documentService.GeneratorHtmlAsync(GetRelativePath(e.FullPath));
+        }
+    }
+
+    private bool IsTemporaryFile(string fileName)
+    {
+        return fileName.EndsWith("~") || fileName.EndsWith(".tmp", StringComparison.OrdinalIgnoreCase);
     }
 
     protected override Task ExecuteAsync(CancellationToken stoppingToken)
